refactor: reorganized the entire application by dividing it into modules
