+++ src/modules/appointments/infra/http/routes/appointments.routes.ts
@@ -1,10 +1,15 @@
 import { Router } from 'express';
 import { startOfHour, parseISO } from 'date-fns';
 import { getCustomRepository } from 'typeorm';
-import AppointmentsRepository from '../repositories/AppointmentsRepository';
-import CreateAppointmentService from '../services/CreateAppointmentService';
+
+import AppointmentsRepository from '@modules/appointments/repositories/AppointmentsRepository';
+import CreateAppointmentService from '@modules/appointments/services/CreateAppointmentService';
+
+import ensuredAuthenticated from '@modules/users/infra/http/middlewares/ensuredAuthenticated';
 
 const appointmentsRouter = Router();
+
+appointmentsRouter.use(ensuredAuthenticated);
 
 appointmentsRouter.get('/', async (req, res) => {
   const appointmentsRepository = getCustomRepository(AppointmentsRepository);
