+++ src/modules/appointments/repositories/AppointmentsRepository.ts
@@ -1,9 +1,9 @@
 import { EntityRepository, Repository } from 'typeorm';
-import Appointment from '../models/Appointment';
+import Appointment from '@modules/appointments/infra/typeorm/entities/Appointment';
 
 @EntityRepository(Appointment)
 class AppointmentsRepository extends Repository<Appointment> {
-  public async findByDate(date: Date): Promise<Appointment> {
+  public async findByDate(date: Date): Promise<Appointment | null> {
     const findAppointment = await this.findOne({ where: { date } });
 
     return findAppointment || null;
