+++ src/modules/appointments/services/CreateAppointmentService.ts
@@ -1,7 +1,9 @@
+import { startOfHour } from 'date-fns';
 import { getCustomRepository } from 'typeorm';
-import Appointment from '../models/Appointment';
+
+import AppError from '@shared/errors/AppError';
+import Appointment from '@modules/appointments/infra/typeorm/entities/Appointment';
 import AppointmentsRepository from '../repositories/AppointmentsRepository';
-import AppError from '../errors/AppError';
 
 interface Request {
   provider_id: string;
@@ -12,13 +14,20 @@ class CreteAppointmentService {
   async execute({ provider_id, date }: Request): Promise<Appointment> {
     const appointmentRepository = getCustomRepository(AppointmentsRepository);
 
-    const findAppointment = await appointmentRepository.findByDate(date);
+    const appointmentDate = startOfHour(date);
+
+    const findAppointment = await appointmentRepository.findByDate(
+      appointmentDate,
+    );
 
     if (findAppointment) {
       throw new AppError('This appointment is already booked');
     }
 
-    const appointment = appointmentRepository.create({ provider_id, date });
+    const appointment = appointmentRepository.create({
+      provider_id,
+      date: appointmentDate,
+    });
 
     await appointmentRepository.save(appointment);
