+++ src/modules/users/services/AuthenticateUserService.ts
@@ -2,10 +2,9 @@ import { getRepository } from 'typeorm';
 import { compare } from 'bcryptjs';
 import { sign } from 'jsonwebtoken';
 
-import User from '../models/User';
-import AppError from '../errors/AppError';
-
-import authconfig from '../config/auth';
+import authconfig from '@config/auth';
+import AppError from '@shared/errors/AppError';
+import User from '@modules/users/infra/typeorm/entities/User';
 
 interface Request {
   email: string;
