+++ src/modules/users/services/UpdateUserAvatarService.ts
@@ -1,10 +1,10 @@
 import { getRepository } from 'typeorm';
 import path from 'path';
 import fs from 'fs';
-import uploadConfig from '../config/upload';
+import AppError from '@shared/errors/AppError';
+import uploadConfig from '@config/upload';
 
-import User from '../models/User';
-import AppError from '../errors/AppError';
+import User from '@modules/users/infra/typeorm/entities/User';
 
 interface Request {
   user_id: string;
