+++ src/shared/infra/http/routes/index.ts
@@ -1,0 +1,12 @@
+import { Router } from 'express';
+import appointmentsRouter from '@modules/appointments/infra/http/routes/appointments.routes';
+import userRouter from '@modules/users/infra/http/routes/users.routes';
+import sessionRouter from '@modules/users/infra/http/routes/sessions.routes';
+
+const routes = Router();
+
+routes.use('/appointments', appointmentsRouter);
+routes.use('/users', userRouter);
+routes.use('/sessions', sessionRouter);
+
+export default routes;
