+++ src/shared/infra/http/server.ts
@@ -2,11 +2,11 @@ import 'reflect-metadata';
 
 import express, { json, Request, Response, NextFunction } from 'express';
 import 'express-async-errors';
-import routes from './routes';
-import uploadConfig from './config/upload';
+import uploadConfig from '@config/upload';
+import AppError from '@shared/errors/AppError';
+import routes from '@shared/infra/http/routes';
 
-import './database';
-import AppError from './errors/AppError';
+import '@shared/infra/typeorm';
 
 const app = express();
